backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.2.3
httpx==0.24.1
pydantic==2.5.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-docx==0.8.11
PyPDF2==3.0.1
python-magic==0.4.27
aiofiles==23.2.1
